RecipeVoter made

+++ facecook/src/Controller/Api/V1/RecipeController.php
@@ -59,13 +59,15 @@ class RecipeController extends AbstractController
             $recipe->setSlug($slugger->slugify($recipe->getTitle()));
 
             // The recipe needs to be associated to a user.
-            $recipe->setUser($this->getUser());
+            //$recipe->setUser($this->getUser());
 
             $em = $this->getDoctrine()->getManager();
             $em->persist($recipe);
             $em->flush();
 
-            return $this->json($recipe, 201, []);
+            return $this->json($recipe, 201, [], [
+                'groups' => ['read', 'browse'],
+            ]);
         }
 
         return $this->json($form->getErrors(true, false)->__toString(), 400);
@@ -76,6 +78,9 @@ class RecipeController extends AbstractController
      */
     public function edit(Recipe $recipe, Request $request, RecipeSlugger $slugger): Response
     {
+        // We'll check if the user has the right to edit.
+        $this->denyAccessUnlessGranted('edit', $recipe);
+
         $form = $this->createForm(RecipeType::class, $recipe, ['csrf_protection' => false]);
 
         $sentData = json_decode($request->getContent(), true);
@@ -86,7 +91,9 @@ class RecipeController extends AbstractController
 
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->json($recipe, 200, []);
+            return $this->json($recipe, 200, [], [
+                'groups' => ['read', 'browse'],
+            ]);
         }
 
         return $this->json($form->getErrors(true, false)->__toString(), 400);
@@ -97,6 +104,8 @@ class RecipeController extends AbstractController
      */
     public function delete (Recipe $recipe): Response
     {
+        $this->denyAccessUnlessGranted('delete', $recipe);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($recipe);
         $em->flush();
